converted alumni page into components and optimized the model page
